+++ app/utils/build-options.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-highcharts/utils/build-options';